Splitt Eu Taxonomy financials e2e tests in two different specs
