+++ dataland-frontend/tests/e2e/specs/eu-taxonomy-financials/index.ts
@@ -3,5 +3,6 @@
  */
 describe("EU Taxonomy for financials Company Page tests", () => {
   require("./DataIntegrity");
-  require("./DataUpload");
+  require("./FormCompletion");
+  require("./FileUploadAndLinking");
 });
